pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.3.1"
description = "Discord + SQLAlchemy 2.0 + FastAPI + Strawberry"
authors = ["TrashRoach <rfsnboris@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

"discord.py" = "^2.3.1"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.19"}
uvicorn = "^0.23.2"
fastapi = "^0.100.1"
strawberry-graphql = "^0.198.0"
asyncpg = "^0.28.0"
fastapi-lifespan-manager = "^0.1.0"

[tool.poetry.group.dev.dependencies]
alembic = "^1.11.1"
mypy = "*"
poetry-dotenv-plugin = "^0.2.0"
black = "^23.7.0"
isort = "^5.12.0"

[tool.poetry.group.test]
[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-postgresql = "^5.0.0"
psycopg = "^3.1.9"

[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test* *Tests"
addopts = "-vv -x"
asyncio_mode = "auto"
pythonpath = ["."]

[tool.mypy]
check_untyped_defs = true
namespace_packages = true
plugins = "sqlalchemy.ext.mypy.plugin"
pretty = true
show_error_context = true
exclude = ".venv"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"